backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
python-multipart
pandas
openpyxl
